+++ test/core/core-test-util.ts
@@ -88,7 +88,6 @@ export const commonArgs: Omit<CoreArgs, "oldTarget" | "src" | "srcCache"> = {
   matcher: "icu",
   srcLng: "en",
   targetLng: "de",
-  overwriteOutdated: true,
 };
 
 export async function translateCoreAssert(